i ran into an issue and had to revert back.
switching between login and registering works fine. I just wanted to apply the same logic to the signup form and add a second step to get the user details in one form. Best thing to do is when redirected to the dashbaord a script should run with a new form that pops up over the dashboard to get this info using a patch method